refactor(admin): flatten to single routed tab bar (#3974)
Split `admin.content.vue` into four routed views (Users, Organizations,
Readiness, Activity) and mount them as children of the `/admin` parent
route. The admin layout now renders a single flat tab bar combining the
built-ins with the `config.admin.tabs` extras, so downstream modules
contribute siblings instead of being stacked under a separate "General"
bar. Global concerns (error banner + mailer warning) moved to the layout
to stay visible across every admin tab.

- Split `admin.content.vue` → `admin.users.view.vue`,
  `admin.organizations.view.vue`, `admin.readiness.view.vue`,
  `admin.activity.view.vue`
- Update `admin.router.js`: `''` redirects to `Admin Users`; each section
  gets a dedicated child route (`users`, `organizations`, `readiness`,
  `activity`) keeping CASL meta propagation
- Update `admin.layout.vue`: merge built-in + extra tabs in a single
  `<v-tabs>` row, longest-prefix active-tab resolution keeps detail
  routes (`/admin/users/:id`) highlighting the parent tab
- Readiness and Activity fetch on `mounted()` instead of via a nested
  tab-model watcher
- Redispatch unit tests across the four new views + refresh router and
  layout specs for the new structure
- Document the change in `MIGRATIONS.md` (URL-level breaking change;
  downstream config unchanged)
