+++ src/modules/admin/views/admin.layout.vue
@@ -1,9 +1,26 @@
 <template>
   <v-container fluid>
     <PageHeader icon="fa-solid fa-user-tie" title="Admin" />
+    <v-alert
+      v-if="error"
+      type="error"
+      variant="tonal"
+      density="compact"
+      closable
+      class="mx-2 mt-2"
+      :class="config.vuetify.theme.rounded"
+      icon="fa-solid fa-circle-exclamation"
+      @click:close="clearError"
+      ><span class="text-body-medium">{{ error }}</span></v-alert
+    >
     <v-tabs v-model="activeTab" color="primary" class="px-2">
-      <v-tab :to="basePath" :value="basePath" exact class="text-none text-body-medium"
-        ><v-icon icon="fa-solid fa-gauge-high" size="small" class="mr-2"></v-icon>General</v-tab
+      <v-tab
+        v-for="tab in builtInTabs"
+        :key="tab.value"
+        :to="tabTo(tab)"
+        :value="tabTo(tab)"
+        class="text-none text-body-medium"
+        ><v-icon :icon="tab.icon" size="small" class="mr-2"></v-icon>{{ tab.label }}</v-tab
       >
       <v-tab
         v-for="extraTab in extraTabs"
@@ -16,6 +33,18 @@
       >
     </v-tabs>
     <router-view />
+    <v-alert
+      v-if="showMailerWarning"
+      type="warning"
+      variant="tonal"
+      density="compact"
+      class="mx-2 mt-4"
+      :class="config.vuetify.theme.rounded"
+      icon="fa-solid fa-triangle-exclamation"
+      ><span class="text-body-medium"
+        >No mailer configured. Users can register with any email without verification. Set up SMTP to enable email verification.</span
+      ></v-alert
+    >
   </v-container>
 </template>
 <script>
@@ -23,28 +52,52 @@
  * Module dependencies.
  */
 import PageHeader from '../../core/components/core.pageHeader.component.vue';
+import { useAdminStore } from '../stores/admin.store';
+import { useAuthStore } from '../../auth/stores/auth.store';
+
+/**
+ * Built-in admin tabs.
+ *
+ * Each entry maps to a routed child of the admin parent route (see
+ * `admin.router.js`). They are rendered inline alongside the extras from
+ * `config.admin.tabs` to form a single flat tab bar.
+ */
+const BUILT_IN_TABS = Object.freeze([
+  { value: 'users', label: 'Users', icon: 'fa-solid fa-users', route: 'users' },
+  { value: 'organizations', label: 'Organizations', icon: 'fa-solid fa-building', route: 'organizations' },
+  { value: 'readiness', label: 'Readiness', icon: 'fa-solid fa-clipboard-check', route: 'readiness' },
+  { value: 'activity', label: 'Activity', icon: 'fa-solid fa-clock-rotate-left', route: 'activity' },
+]);
 
 /**
  * Component definition.
  *
  * `admin.layout.vue` is the parent layout component for the admin section.
- * It renders the page header plus the top-level tab bar (General + extra
- * tabs from `config.admin.tabs`) and a `<router-view>` that hosts:
+ * It renders the page header, a single flat tab bar (built-in Users /
+ * Organizations / Readiness / Activity + extras from `config.admin.tabs`)
+ * and a `<router-view>` that hosts the active child:
  *
- *  - The base "General" route → `admin.content.vue` (Users / Orgs / ...)
- *  - User / Organization detail views (`/admin/users/:id`, ...)
+ *  - Built-in tabs (`/admin/users`, `/admin/organizations`, …) are routed
+ *    children defined in `admin.router.js`.
+ *  - User / Organization detail views (`/admin/users/:id`, …) are also
+ *    children — they deep-link inside the layout.
  *  - Any child route injected via `injectAdminChildren` (downstream tabs)
+ *    renders through the same `<router-view>`.
  *
  * Extra tabs are config-driven; their `route` may be relative (preferred,
  * e.g. `'knowledge'`) or a legacy absolute path (`'/admin/knowledge'`).
  * Both are supported — relative is resolved against `/admin/`.
+ *
+ * Global concerns (error banner, mailer warning) are rendered here so
+ * they stay visible across all admin tabs, including downstream extras.
  */
 export default {
   name: 'AdminLayout',
   components: { PageHeader },
   data() {
     return {
-      activeTab: '/admin',
+      activeTab: '/admin/users',
+      builtInTabs: BUILT_IN_TABS,
     };
   },
   computed: {
@@ -55,6 +108,20 @@ export default {
      */
     basePath() {
       return '/admin';
+    },
+    /**
+     * @desc Global error from the admin store (surfaced on any admin tab).
+     * @returns {string|null}
+     */
+    error() {
+      return useAdminStore().error;
+    },
+    /**
+     * @desc Whether to show the mailer warning across admin tabs.
+     * @returns {boolean}
+     */
+    showMailerWarning() {
+      return useAuthStore().serverConfig?.mail?.configured === false;
     },
     /**
      * @desc Returns validated extra admin tabs from `config.admin.tabs`.
@@ -92,6 +159,12 @@ export default {
     },
   },
   methods: {
+    /**
+     * @desc Clear the global admin error banner.
+     */
+    clearError() {
+      useAdminStore().error = null;
+    },
     /**
      * @desc Validate a single tab descriptor (shared by `extraTabs`).
      * @param {unknown} tab - Raw entry from `config.admin.tabs`.
@@ -138,7 +211,7 @@ export default {
     /**
      * @desc Resolve a tab descriptor to a concrete absolute path.
      *       Relative paths are joined under `/admin/`.
-     * @param {{ route: string }} tab - The extra tab descriptor.
+     * @param {{ route: string }} tab - The tab descriptor.
      * @returns {string} Absolute path the `<v-tab>` should link to.
      */
     tabTo(tab) {
@@ -147,18 +220,18 @@ export default {
     },
     /**
      * @desc Map a router path to the `v-tab` value it should activate.
-     *       Exact-match for the base path, then prefix-match for extra
-     *       tabs. Falls back to the base path so the indicator never
-     *       leaves a visible tab empty.
+     *       Prefix-match across built-in + extra tabs, longest-match wins
+     *       so detail routes (e.g. `/admin/users/abc`) still light up the
+     *       parent tab (`/admin/users`).
      * @param {string} path - Current route path.
-     * @returns {string} The matching tab's `value`.
+     * @returns {string} The matching tab's `value` (absolute path).
      */
     resolveActiveTab(path) {
-      if (path === this.basePath) return this.basePath;
-      const match = this.extraTabs
-        .map((tab) => this.tabTo(tab))
-        .find((to) => path === to || path.startsWith(`${to}/`));
-      return match || this.basePath;
+      const candidates = [...this.builtInTabs, ...this.extraTabs].map((tab) => this.tabTo(tab));
+      const match = candidates
+        .filter((to) => path === to || path.startsWith(`${to}/`))
+        .sort((a, b) => b.length - a.length)[0];
+      return match || `${this.basePath}/users`;
     },
   },
 };
